Reorganize project files in new directories
