+++ screens/initial_screen.py
@@ -1,4 +1,4 @@
-from screen import Button, Screen
+from screens.screen import Button, Screen
 from config import CONFIG_SCREEN, GAME_SCREEN, INITIAL_SCREEN
 import pygame
